[#86] Pin pyinstaller version used for container image build
The pyinstaller version has been pinned to 5.9.0 to prevent usage
of the recently released 5.10.0 version which contains an internal
API change that causes staticx to break.

See https://github.com/JonathonReinhart/staticx/issues/235

Also changed the Dockerfile to install the standard patchelf OS package
instead of using pip install because this is the approach recommended
by the staticx installation guide.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>