[#122] Warn on DBC signal name usage across multiple messages
During startup, the DBC Feeder reads in CAN message definitions
from arbitrary database files. When using the same signal name in
different message definitions, care needs to be taken that the semantics
of the signal are the same across all messages.

The DBC Feeder has been changed to detect such cases and issue a
corresponding warning if the signal is being mapped to a VSS datapoint.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>